+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 #########
 
+0.2.1 - Under development
+-------------------------
+
+* Add an Affiliation model and associates it to the Person abstract model through a ManyToManyField
+* Update the ``import_journals_from_fedora`` management command to import affiliations
+
 0.2.0
 -----